+++ src/app/app.html
@@ -1,1 +1,2 @@
 <app-home-pages />
+<app-footer />
