+++ src/app/components/card/card.component.html
@@ -55,5 +55,4 @@
       Ver {{ showMore ? "menos" : "más" }}
     </button>
   }
-  <app-footer />
 </main>
