+++ src/app/components/modal-dates/edit-modal/modal-editHero.component.ts
@@ -2,7 +2,7 @@ import { Component, inject } from '@angular/core';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ModalEditService } from '@services/core/modal-services/modal-edit.service';
 import { FormComponent } from '@form/form.component';
-
+import { CoreModalServices } from '@core/modal-services/coreModal.service';
 @Component({
   selector: 'app-modal-edit-hero',
   templateUrl: './edit-modalHero.html',
@@ -11,8 +11,8 @@ import { FormComponent } from '@form/form.component';
 })
 export class ModalEditHeroComponent {
   public readonly modalEditService = inject(ModalEditService);
+  public readonly coreModal = inject(CoreModalServices);
   public readonly snackBar = inject(MatSnackBar);
-  public readonly formControlUpdate = this.modalEditService.formControlUpdate;
 
   onSubmit() {
     this.modalEditService.onSubmit();
